Test that get_* functions don't create directories
According to the documentation, the get_* functions should not touch the
file system while the functions without the get_ prefix create the
returned directory (and all parent directories).

But with the current implementation, the get_* functions do create
the data and app root directories if they don’t exist, see bug report
34 in the original repository:
	https://github.com/AndyBarron/app-dirs-rs/issues/34

This patch adds a test suite that demonstrates the problem.