android: Pass directory "type" string when calling `getExternalFilesDir`
My Android 12 phone throws the following error when trying to resolve a
path with `SharedData` or `SharedConfig`:

    E AndroidRuntime: java.lang.NoSuchMethodError: no non-static method "Landroid/app/NativeActivity;.getExternalMediaDir()Ljava/io/File;"

As it turns out the [getExternalFilesDir] function takes a `String`
argument with type of directory (ie. pictures, music, video or
downloads) to get, or `null` for the root.  This `null`/`Void` argument
is trivially passed through `JNI`.

It is uncertain whether this function was previously available without
arguments or tested at all, as it seems unlikely for Android to make
such a breaking function-signature change.

Also remove the deprecation comment.  This function is still well in use
and not listed as deprecated; the only function deprecated in Android 11
(API level 30) is [getExternalMediaDirs], which is not used here anyway.

getExternalFilesDir: https://developer.android.com/reference/android/content/Context#getExternalFilesDir(java.lang.String)
getExternalMediaDirs: https://developer.android.com/reference/android/content/Context#getExternalMediaDirs()